# BookStore/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/addBookToCartByBookId": {
            "post": {
                "summary": "Add Book to cart",
                "description": "User Can Add Book to cart ",
                "operationId": "e06accabe2f8708ced7cea624175038f",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "book_id"
                                ],
                                "properties": {
                                    "book_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Book added to Cart Sucessfully"
                    },
                    "404": {
                        "description": "Invalid authorization token"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/deleteBookByCartId": {
            "post": {
                "summary": "Delete the book from cart",
                "description": " Delete cart ",
                "operationId": "ddf2bd9696caf3ef09cffc5794f1a78d",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Book deleted Sucessfully from cart"
                    },
                    "404": {
                        "description": "Invalid authorization token"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/getAllBooksByUserId": {
            "get": {
                "summary": "Get All Books Present in Cart",
                "description": " Get All Books Present in Cart ",
                "operationId": "d25cc9ca2acf117c491545cbb31a5ce8",
                "requestBody": {},
                "responses": {
                    "404": {
                        "description": "Invalid authorization token"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/addAddress": {
            "post": {
                "summary": "Add Address",
                "description": "User Can Add Address ",
                "operationId": "613f059314af172e73cf474da459f525",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "address",
                                    "city",
                                    "state",
                                    "landmark",
                                    "pincode",
                                    "address_type"
                                ],
                                "properties": {
                                    "address": {
                                        "type": "string"
                                    },
                                    "city": {
                                        "type": "string"
                                    },
                                    "state": {
                                        "type": "string"
                                    },
                                    "landmark": {
                                        "type": "string"
                                    },
                                    "pincode": {
                                        "type": "integer"
                                    },
                                    "address_type": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Address Added Successfully"
                    },
                    "401": {
                        "description": "Address alredy present for the user"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/updateBookQuantityInCart": {
            "post": {
                "summary": "Add Quantity to Existing Book in cart",
                "description": " Add Book Quantity  in cart",
                "operationId": "9611d240fc80dc0dd0326525f51f56b6",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "id",
                                    "book_quantity"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    },
                                    "book_quantity": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Book Quantity updated Successfully"
                    },
                    "404": {
                        "description": "Invalid authorization token"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/placeOrder": {
            "post": {
                "summary": "Place  Order",
                "description": "Place a Order",
                "operationId": "a0ea5f9c62cd4c24a6ddabaa99aa6c14",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "address_id",
                                    "cart_id"
                                ],
                                "properties": {
                                    "address_id": {
                                        "type": "integer"
                                    },
                                    "cart_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order Placed Successfully"
                    },
                    "404": {
                        "description": "Address Not Found"
                    },
                    "401": {
                        "description": "Invalid Authorization Token"
                    },
                    "409": {
                        "description": "Already Placed an Order"
                    },
                    "406": {
                        "description": "Book Stock is Not Available in The Store"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "description": "Laravel bearer token authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}